requirements.txt
affine==2.4.0
amqp==5.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
attrs==24.1.0
billiard==4.2.0
bqplot==0.12.43
branca==0.7.2
cachetools==5.4.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
colour==0.1.5
comm==0.2.2
contourpy==1.2.0
crispy-bootstrap5==2023.10
cron-descriptor==1.4.3
cryptography==42.0.1
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
Django==5.0.1
django-allauth==0.60.1
django-bitfield==2.2.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-geo==0.8
django-js-asset==2.2.0
django-mptt==0.16.0
django-rest-framework==0.1.0
django-separatedvaluesfield==0.6.0
django-simple-history==3.4.0
django-tenant-schemas==1.12.0
django-tenants==3.6.1
django-tenants-celery-beat==0.2.1
django-timezone-field==6.1.0
django-uuidfield==0.5.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
drf-yasg==1.21.7
earthengine-api==0.1.414
eerepr==0.0.4
executing==2.0.1
folium==0.17.0
fonttools==4.50.0
future==1.0.0
GDAL==3.9.1
geemap==0.33.1
geocoder==1.38.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.6
inflection==0.5.1
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipyleaflet==0.19.2
ipython==8.26.0
ipytree==0.2.2
ipywidgets==8.1.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter-leaflet==0.19.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
kombu==5.3.5
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.7
numpy==1.26.4
oauthlib==3.2.2
openeo==0.27.0
ordered-set==4.1.0
packaging==23.2
pandas==2.2.0
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
plotly==5.23.0
postgis==1.0.4
proj==0.2.0
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==5.27.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
pyperclip==1.9.0
pyshp==2.3.1
python-box==7.2.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
rasterio==1.3.10
ratelim==0.1.6
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.0
scooby==0.10.0
setuptools==68.2.2
shapely==2.0.3
six==1.16.0
snuggs==1.4.7
South==1.0.2
sqlparse==0.4.4
stack-data==0.6.3
tenacity==9.0.0
tenant-schemas-celery==2.2.0
threadpoolctl==3.5.0
traitlets==5.14.3
traittypes==0.2.1
types-python-dateutil==2.8.19.20240106
tzdata==2023.4
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.41.2
widgetsnbextension==4.0.11
wrapt==1.16.0
xarray==2024.1.1
xyzservices==2024.6.0
